match stats, player profiles, 128-slot support; bump 0.70.7
Ports WORR's match-statistics and player-profile systems: persistent per-gametype Elo ratings, structured JSON match exports with HTML reports and atomic catalog, bounded departure/forfeit settlement, and reconnect-aware profile reconciliation.

Raises the connected-client ceiling from 32 to 128 (MAX_LOBBY_PLAYERS), normalizing maxclients during PreInit so the engine slab and game allocation cannot diverge.

Hardens the release workflow: all third-party actions commit-pinned, Copilot CLI verified via SHA-512 tarball, build receipt binding source commit and DLL hash, provenance JSON + SHA256SUMS covering every asset, and verified-tag publication.

Adds check-release-workflow.ps1, check-tooling-contracts.ps1, run-fuzz-smoke.ps1, clang-tidy-blocking-checks.txt, and msvc-analyze-blocking-warnings.txt. Fuzz build now stages ASan runtime and executes a bounded corpus smoke run. Cppcheck is blocking in CI and locally.

Fixes entity lifetime safety across damage callbacks, physics, turrets, doppelgangers, spheres, trains, projectiles, and the savegame loader. Adds mm_ordnance_identity.h generation/arena stamps for all delayed ordnance. Menus now capture movement for live players, not just spectators.

+++ src/sgame/monsters/m_supertank.cpp
@@ -235,8 +235,8 @@ static void supertankGrenade(gentity_t *self) {
 	vec3_t aim_point;
 	PredictAim(self, self->enemy, start, 0, false, crandom_open() * 0.1f, &forward, &aim_point);
 
-	for (float speed = 500.f; speed < 1000.f; speed += 100.f) {
-		if (!M_CalculatePitchToFire(self, aim_point, start, forward, speed, 2.5f, true))
+	for (int32_t speed = 500; speed < 1000; speed += 100) {
+		if (!M_CalculatePitchToFire(self, aim_point, start, forward, static_cast<float>(speed), 2.5f, true))
 			continue;
 
 		monster_fire_grenade(self, start, forward, 50, speed, flash_number, 0.f, 0.f);
